app fully working after breaking down frontend components
